+++ NEWS.md
@@ -51,6 +51,8 @@
 
 * `select()` and `vars()` now treat `NULL` as empty inputs (#3023).
 
+* Fix `summarise()` for empty data frames with zero columns (#3071).
+
 * Add error for `distinct()` if any of the selected columns are of type `list` (#3088, @foo-bar-baz-qux).
 
 * `sample_n()` and `sample_frac()` on grouped data frame are now faster especially for those with large number of groups (#3193, @saurfang).